Migra el código dentro de `Movement.cs` a un State Pattern
