+++ meica4/Assets/DoubleJumpItem.cs
@@ -4,13 +4,13 @@ public class DoubleJumpItem : MonoBehaviour
 {
     private void OnTriggerEnter2D(Collider2D other)
     {
-        if (other.CompareTag("Player")) // Asegúrate de que el jugador tenga el tag "Player"
+        if (other.CompareTag("Player")) // Asegï¿½rate de que el jugador tenga el tag "Player"
         {
-            Movimiento player = other.GetComponent<Movimiento>();
+            Movement player = other.GetComponent<Movement>();
             if (player != null)
             {
                 player.EnableDoubleJump(); // Activa el doble salto
-                Destroy(gameObject); // Destruye el ítem después de recogerlo
+                Destroy(gameObject); // Destruye el ï¿½tem despuï¿½s de recogerlo
             }
         }
     }
