+++ meica4/Assets/oak_woods_v1.0/scripts/enemy/enemy.cs
@@ -102,7 +102,7 @@ public class EnemyAI : MonoBehaviour
         {
             if (Time.time > lastAttackTime + timeBetweenAttacks)
             {
-                Movimiento playerLife = other.GetComponent<Movimiento>();
+                Movement playerLife = other.GetComponent<Movement>();
                 if (playerLife != null)
                 {
                     Debug.Log("⚔️ Enemigo ha golpeado al jugador por " + damage + " de daño.");
